Continuing development

+++ plumbing/ps_modules/common/ps_common_typedefs.h
@@ -10,6 +10,6 @@
 #define ps_common_typedefs_h
 
 typedef uint8_t ps_packet_type_t;
-typedef uint16_t ps_packet_source_t;
+typedef uint8_t ps_packet_source_t;
 
 #endif /* ps_common_typedefs_h */
